pyproject.toml
[tool.black]
line-length = 130
target-version = ['py38']
skip-string-normalization = true

[tool.poetry]
name = "moneyholic"
version = "0.1.0"
description = ""
authors = ["adnasal <adnasalkovic@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytz="==2021.1"
Django="==3.2.12"
gunicorn="==20.1.0"
newrelic="==6.4.0.157"
django-dotenv="==1.4.2"
django-3-jet="==1.0.8"
celery="==4.4.7"
virtualenv="==20.16.4"
psycopg2-binary="==2.9.3"
redis="==3.5.3"
django-model-utils="==4.1.1"
django_unique_upload="==0.2.1"
django-summernote="==0.8.11.6"
django-celery-beat="==2.0.0"
flower="==0.9.7"
django-activity-stream="==0.10.0"
django-money="==2.0"
djangorestframework="==3.12.4"
djangorestframework-simplejwt="==4.7.1"
Markdown="==3.3.4"
beautifulsoup4="==4.10.0"
lxml="==4.9.1"
drf-yasg="==1.20.0"
django-filter="==2.4.0"
django-cors-headers="==3.7.0"
django-rest-passwordreset="==1.1.0"
django-rest-swagger="==2.2.0"
easy-thumbnails="==2.7.1"
django-auto-prefetching="==0.1.10"
social-auth-core="==4.1.0"
social-auth-app-django="==4.0.0"
ipython="==7.24.0"
sentry-sdk="==1.1.0"
django-inlinecss="==0.3.0"
django-storages="==1.11.1"
boto3="==1.17.84"
django-health-check="==3.16.4"

[tool.poetry.dev-dependencies]
ipdb="==0.13.8"
black="==21.9b0"
django-extensions="==3.1.3"
pre-commit="==2.15.0"
factory-boy="==3.2.0"
django-nose="==1.4.7"
coverage="==5.5"
setuptools="==58"
unittest-prettify="==1.0.post4"
locust="==2.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
